+++ packages/eslint-plugin/CHANGELOG.md
@@ -1,3 +1,6 @@
+# Next
+- Added `jsx-a11y/anchor-is-valid` to enforce all anchors to be valid, navigable elements.
+
 # 0.0.3
 - Added rules to follow our guides related to strings.